+++ samples/spring-fibonacci-action/src/main/java/com/example/fibonacci/LimitedFibonacciAction.java
@@ -3,7 +3,7 @@ package com.example.fibonacci;
 import io.grpc.Status;
 import kalix.javasdk.action.Action;
 import kalix.javasdk.action.ActionCreationContext;
-import kalix.springsdk.impl.KalixClient;
+import kalix.springsdk.KalixClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
