+++ sdk/spring-sdk/src/it/java/com/example/wiring/actions/echo/ShortenedEchoAction.java
@@ -18,7 +18,7 @@ package com.example.wiring.actions.echo;
 
 import kalix.javasdk.action.Action;
 import kalix.javasdk.action.ActionCreationContext;
-import kalix.springsdk.impl.KalixClient;
+import kalix.springsdk.KalixClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
