+++ sdk/spring-sdk/src/main/scala/kalix/springsdk/KalixClient.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 Lightbend Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package kalix.springsdk
+import akka.annotation.ApiMayChange
+import kalix.springsdk.impl.RestDeferredCall
+
+trait KalixClient {
+
+  @ApiMayChange
+  def post[P, R](uri: String, body: P, returnType: Class[R]): RestDeferredCall[P, R]
+
+  @ApiMayChange
+  def get[R](uri: String, returnType: Class[R]): RestDeferredCall[_, R]
+}
